feat(firmware): production build target (ADR-0040 steps 1-2)
Terminals ran the dev firmware, which blocks up to 10s at every boot waiting for
a USB serial console. Add a production build that skips that wait.

- Extract the whole application into //maco_firmware/apps/app_main as
  RunApp(AppConfig); apps/dev/main.cc becomes a thin shim. Behavior-preserving:
  dev still passes wait_for_usb_serial=true. AppConfig is passed via a file-scope
  static (not a captured lambda) so system::Init() still takes a plain function
  pointer — a lambda capturing AppConfig can exceed pw::Function's inline
  capacity on the 32-bit target.
- apps/prod builds the same library with wait_for_usb_serial=false. Adds the
  `build:prod` config (-c opt), `./pw prod-flash` / `prod-console` (both pass
  --config=prod so the flashed image and console token database match), and adds
  prod to `./pw build p2`.

Verified: host + p2 builds pass; prod builds in both fastbuild and --config=prod.
Watchdog (enable_watchdog) stays off until step 4. prod memory_budget deferred
until a golden can be generated from a real build.
