+++ pw
@@ -321,6 +321,86 @@ case "$1" in
       --device "$DEVICE" \
       --device-serial-suffix "$SERIAL_SUFFIX"
     ;;
+  prod-flash)
+    # Flash the optimized production firmware (ADR-0040). --config=prod builds
+    # -c opt; prod boots without the 10s serial wait and arms the watchdog.
+    echo "🔌 Flashing PRODUCTION firmware (optimized)..."
+    exec bazelisk run --config=prod //maco_firmware/apps/prod:flash
+    ;;
+  prod-console)
+    # Connect console to a prod-flashed device. Builds the prod ELF with
+    # --config=prod so the token database matches the flashed (opt) image.
+    # Usage: ./pw prod-console [--device-serial=XXXX]
+
+    DEVICE_SERIAL=""
+    for arg in "$@"; do
+      case "$arg" in
+        --device-serial=*)
+          DEVICE_SERIAL="${arg#*=}"
+          ;;
+      esac
+    done
+
+    find_particle_device() {
+      local serial_filter="$1"
+      local timeout=30
+      local elapsed=0
+
+      while (( elapsed < timeout )); do
+        local devices=(/dev/particle_*)
+
+        if [[ -e "${devices[0]}" ]]; then
+          if [[ -n "$serial_filter" ]]; then
+            for dev in "${devices[@]}"; do
+              if [[ "$dev" == *"$serial_filter" ]]; then
+                echo "$dev"
+                return 0
+              fi
+            done
+          elif (( ${#devices[@]} == 1 )); then
+            echo "${devices[0]}"
+            return 0
+          else
+            echo "🔌 Multiple Particle devices found:" >&2
+            local i=1
+            for dev in "${devices[@]}"; do
+              echo "   $i) $dev" >&2
+              ((i++))
+            done
+            echo -n "Select device [1-${#devices[@]}]: " >&2
+            read -r selection
+            if [[ "$selection" =~ ^[0-9]+$ ]] && (( selection >= 1 && selection <= ${#devices[@]} )); then
+              echo "${devices[$((selection-1))]}"
+              return 0
+            fi
+            echo "❌ Invalid selection" >&2
+            return 1
+          fi
+        fi
+
+        if (( elapsed == 0 )); then
+          if [[ -n "$serial_filter" ]]; then
+            echo "⏳ Waiting for device *$serial_filter..." >&2
+          else
+            echo "⏳ Waiting for Particle device..." >&2
+          fi
+        fi
+        sleep 1
+        ((elapsed++))
+      done
+
+      echo "❌ No matching Particle device found" >&2
+      return 1
+    }
+
+    DEVICE=$(find_particle_device "$DEVICE_SERIAL") || exit 1
+    SERIAL_SUFFIX="${DEVICE##*/dev/particle_}"
+
+    echo "🔌 Connecting to PROD $DEVICE (serial: ...$SERIAL_SUFFIX)..."
+    MACO_PROJECT_ROOT="$_PW_SCRIPT_DIR" exec bazelisk run --config=prod //maco_firmware/apps/prod:console -- \
+      --device "$DEVICE" \
+      --device-serial-suffix "$SERIAL_SUFFIX"
+    ;;
 esac
 
 # For all other commands, use the workflows launcher
